modules/embeddings.py
import os
from typing import List

from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()

EMBEDDING_MODEL = "text-embedding-3-small"

client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))


def embed_text(text: str) -> List[float]:
    """
    Generate an embedding vector for a single text string.
    """
    response = client.embeddings.create(
        model=EMBEDDING_MODEL,
        input=text
    )
    return response.data[0].embedding


def embed_chunks(chunks: List[str]) -> List[List[float]]:
    """
    Generate embeddings for a list of text chunks.
    """
    if not chunks:
        return []

    response = client.embeddings.create(
        model=EMBEDDING_MODEL,
        input=chunks
    )

    return [item.embedding for item in response.data]