requirements.txt
streamlit
boto3
python-dotenv
pypdf
python-docx
tiktoken
chromadb
openai
reportlab
tenacity